Extract #select_entries (and manual selection flag) into new SelectionHelper
